+++ content/lib/pyswitch/clients/local/actions/custom.py
@@ -1,0 +1,71 @@
+from ....controller.callbacks import Callback
+from ....controller.actions import Action
+from ....colors import Colors
+from adafruit_midi.midi_message import MIDIMessage
+
+# Sends a raw, arbitrary MIDI message.
+def CUSTOM_MESSAGE(message,                 # Raw MIDI message bytes (as list, for example [176, 80, 0] for a control change)
+                   message_release = None,  # Raw MIDI message to be sent on releasing the button (default: None)
+                   text = "",
+                   color = Colors.WHITE,
+                   led_brightness = 0.15,   # LED brighness in range [0..1]
+                   display = None,
+                   use_leds = True,
+                   id = None,
+                   enable_callback = None
+    ):
+    return Action({
+        "callback": _CustomMessageCallback(
+            message = message,
+            message_release = message_release,
+            color = color,
+            led_brightness = led_brightness,
+            text = text
+        ),
+        "display": display,
+        "useSwitchLeds": use_leds,
+        "id": id,
+        "enableCallback": enable_callback
+    })            
+
+
+class _CustomMessageCallback(Callback):
+    class _RawMessage(MIDIMessage):
+        def __init__(self, data):
+            self.__data = bytearray(data)
+
+        def __bytes__(self):
+            return self.__data
+
+    def __init__(self, 
+                 message,
+                 message_release,
+                 color,
+                 led_brightness,
+                 text
+        ):
+        super().__init__()
+        
+        self.__message = message
+        self.__message_release = message_release
+        self.__color = color
+        self.__text = text
+        self.__led_brightness = led_brightness
+
+    def init(self, appl, listener = None):
+        self.__appl = appl
+
+    def push(self):
+        self.__appl.client.midi.send(self._RawMessage(self.__message))
+
+    def release(self):
+        if self.__message_release:
+            self.__appl.client.midi.send(self._RawMessage(self.__message_release))
+    
+    def update_displays(self):
+        self.action.switch_color = self.__color
+        self.action.switch_brightness = self.__led_brightness
+
+        if self.action.label:
+            self.action.label.text = self.__text
+            self.action.label.back_color = self.__color