+++ web/htdocs/clients/kemper/virtual/VirtualKemperParameters.js
@@ -13,17 +13,18 @@ class VirtualKemperParameters {
     /**
      * Tries to return a meaningful message name
      */
-    getMessageName(message) {
+    getMessageProperties(message) {
         for (const [key, entry] of this.#data) {
             // Try to parse the message
-            const parseResult = entry.parse(message, true)
+            let parseResult = entry.parse(message, true)
             if (parseResult) {
-                return ((typeof parseResult == "string") ? (parseResult + " ") : "") + entry.getDisplayName();
+                return parseResult;
             }
 
             // Check if the message could be sent from the parameter
-            if (entry.parseSendMessage(message)) {
-                return entry.getDisplayName();
+            parseResult = entry.parseSendMessage(message);
+            if (parseResult) {
+                return parseResult;
             }
         }