+++ web/htdocs/clients/kemper/virtual/VirtualKemperProtocol.js
@@ -27,11 +27,17 @@ class VirtualKemperProtocol {
     /**
      * Tries to return a meaningful message name
      */
-    getMessageName(message) {
+    getMessageProperties(message) {
         const keppAliveMsg = [240, 0, 32, 51, this.#client.options.productType, 127, 126, 0, 127]
-        if (Tools.compareArrays(keppAliveMsg, message.slice(0, keppAliveMsg.length))) return "Protocol KeepAlive";
+        if (Tools.compareArrays(keppAliveMsg, message.slice(0, keppAliveMsg.length))) return {
+            name: "Protocol KeepAlive",
+            value: ""
+        };
 
-        if (this.parse(message, true)) return "Protocol Init"
+        if (this.parse(message, true)) return {
+            name: "Protocol Init",
+            value: ""
+        };
     }
 
     /**