+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     
     - Added Parameter Up/Down action: This can for example be used to lower/raise the value of a parameter by a given offset. (Also for this there are several value display possibilities). This can replace missing expression pedals :) I personally use it for changing the Reverb Mix on the fly without sacrificing my Volume Pedal. See the "Parameter Up Down" example for the Nano 4 for demonstration.
 
+    - Added a custom MIDI message action. This (for now) just sends any arbitrary byte sequence set to the message parameter. Especially useful when exploring the MIDI specification of devices like the Kemper, in conjunction with the new MIDI monitor (see Emulator changes).
+
 - Optimizations:
     - Modularized colors out of misc.py (some RAM gain). This import path for color stuff has changed to "pyswitch.colors", if you use that in your config somewhere, it has to be adjusted. All examples are already adapted.
     - Changed mapping names to fit better into the preview displays
@@ -15,7 +17,8 @@
     - Renamed mapping MAPPING_DELAY_MIX to MAPPING_DLY_REV_MIX (config needs to be changed only if you have this mapping adressed in the "Other" action or by code)
 
 ### Emulator 2.4.2.9
-    - Added an optional MIDI Traffic Monitor, showing all MIDI messages sent and received by the PySwitch emulator.
+    - Added a MIDI Traffic Monitor, showing all MIDI messages sent and received by the PySwitch emulator.
+    - Other cosmetic optimizations
 
 # PySwitch v2.4.1
 - Optimizations for performance/RAM: